requirements.txt
# Travel Business RAG WhatsApp Bot - Dependencies
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Production WSGI server
gunicorn>=21.0.0

# File uploads support
python-multipart>=0.0.6

# HTTP client
httpx==0.27.2
requests>=2.31.0

# LLM APIs
openai==1.54.4

# LangChain - Session-based conversational memory
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Vector math & tokenization
numpy==1.26.4
tiktoken>=0.5.0

# Knowledge Graph
networkx>=3.2

# Hybrid Retrieval (BM25)
rank-bm25>=0.2.2

# Data Models
pydantic>=2.5.0

# Environment variables
python-dotenv==1.0.1

# Optional: Redis for production session storage (uncomment if needed)
# redis>=5.0.0

# Optional: Cross-Encoder Reranking (uncomment if needed)
# sentence-transformers>=2.2.0
# torch>=2.0.0

# Optional: YouTube Data API v3 (set YOUTUBE_API_KEY env var)
# If not set, YouTube search falls back to predefined video URLs
